feat: Explicitly close connections on end
To avoid the overhead of reinitializing HTTP connections when using an
HTTP PIP resolver, we use an HTTP agent with connection keepAlive set to
true.

However, avoiding explicitly closing these connections uses extra
resources on the server (the server will assume the connection is open
after the resolver is done with it's work). This effect is minor but we
might as well fix it.

A more important effect of this change is that it drastically speeds up
our tests! Previously, the HTTP server we created for the tests would
wait for a full 2 minutes after the tests were done before closing the
connection and shutting down. Now that the connection is explicitly
closed by the resolver, the Node process can exit immediately after the
last test is finished! 🏇

See the HTTP [agent.destroy()](https://nodejs.org/api/http.html#http_agent_destroy)
docs for more info.

Fixes #42