feat: Expose localPipResolver function
In some cases, such as running the PIP service itself, we don't want
wof-admin-lookup to autodetect whether to to use a remote or local PIP
service.

This allows us to keep a single `pelias.json` for an entire Pelias
deployment, with remote PIP turned on, without the PIP service
attempting to recursively connect to itself and break.