fix(lookupStream): properly use lookupStream end event
The pip resolver was being shut down on the Node.js stream 'finish'
event, rather than the 'end' event. The finish event is called after all
data is _read_ into the stream, but there may be some processing left to
do.

The end result was that some records could never be processed and code
using this stream would hang.