feat(worker): pass WOF summary file in message instead of via file
Each admin lookup worker generates summary information with information
about the WOF hierarchy, which is passed to the main process. Currently, this
information is passed by writing to a file in a temporary location.

This isn't desirable as the code may be running in a Docker container
that has a read-only filesystem, and in general it's not nice to write
random files anyway.

By passing the same data back in the message body of the "loaded" IPC
message sent back to the main process, this temporary file can be
avoided.

Presumably, this method was avoided because passing what could be
many megabytes of JSON between processes in this way was scary. However,
even when testing with a full planet admin lookup setup, there was no
noticeable difference with this new method.

Connects https://github.com/pelias/pelias/issues/745
Fixes https://github.com/pelias/polylines/issues/146