Restore Australian data
This change restores the data for Australia that was removed in
https://github.com/pelias/wof-admin-lookup/pull/281.

Its not quite a pure reversion, as the data has also been updated with
the latest postal cities run, as the rest of the data was in
https://github.com/pelias/wof-admin-lookup/pull/279.

Australian postal cities data is still not _used_ to determine possible
localities. However now that decision is made in the proper place in the
code, rather than by removing the data.