fix(postalCities): Only change locality if borough is _not_ changed
Because we now manage postal cities replacements for both borough and
locality values, we have to be a bit careful about when we replace each
of them.

After invesitating what looked like an error in our logic for
determining the best postal cities match in
https://github.com/pelias/wof-admin-lookup/issues/288, it turns out a
significant issue we were seeing in Brooklyn was actually a more subtle
bugs.

In short, it looks like if we replace the borough field with with a
postal city value, we should _not_ replace the localtiy value. This PR
implements that change.