Enable postalcode lookup
This diff enables lookup in the `postalcode` layer, guarded by a config flag.
Because the postal code data can amount to a significant size (12.6 GB as of
today), this setting is disabled by default.

The new `lookupPostalCode` setting affects only the "local resolver" mode. If
it's set to true, it will start the `postalcode` layer worker. The layer is
considered "untrusted", just like the `neighbourhood` layer.